reorganize config subpackage: - move most code to new main.py - move specific handlers from start() to separate functions (possibly exaggerating a bit) - add [ui] section support to config reading
